Implement id() and improve name() for ELF
For ELF special handling must be implemented for the main executable
which does not have a name that can be discovered.  This implementation
now uses `env::current_exe` to handle this.  Additionally this now
implements `id()` by loading the ELF binary referenced by a shared
library reference to discover the embedded gnu build id.

This does change behavior quite a bit and the elf code path has been
feature flagged in case developers are not interested in ID handling
which is primarily useful for server side symbolication.

This also removes `Clone` and `Copy` from the shared library trait
on linux. Copy is no longer possible with this implementation and
Clone is not implemented for macos.  This now makes this consistent.